backend/requirements.txt
amqp==5.0.9
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
braintree==4.13.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-bootstrap4==21.1
djangorestframework==3.13.1
dnspython==2.1.0
greenlet==1.1.2
idna==3.3
kombu==5.2.3
Pillow==8.4.0
prompt-toolkit==3.0.24
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
